@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "description": "A GitHub action that creates a seasonal Jekyll post from data files",
   "main": "dist/index.js",
+  "type": "module",
   "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint",
+    "pretest": "eslint .",
+    "test": "jest",
     "pretty-quick": "pretty-quick",
     "build": "ncc build src/index.js && node src/docs.js",
     "prepare": "husky install"
@@ -13,19 +14,20 @@
   "author": "Katy DeCorah",
   "license": "MIT",
   "devDependencies": {
-    "@actions/core": "^1.5.0",
-    "@babel/eslint-parser": "^7.15.7",
-    "@zeit/ncc": "^0.22.3",
-    "eslint": "^7.20.0",
+    "@actions/core": "^1.6.0",
+    "@babel/eslint-parser": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
+    "@vercel/ncc": "^0.33.1",
+    "eslint": "^8.5.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^7.0.2",
-    "jest": "^27.2.2",
+    "husky": "^7.0.4",
+    "jest": "^27.4.5",
     "js-yaml": "^4.1.0",
-    "octokit": "^1.6.1",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0"
+    "octokit": "^1.7.1",
+    "prettier": "^2.5.1",
+    "pretty-quick": "^3.1.2"
   },
   "repository": {
     "type": "git",
